+++ R/unifont.R
@@ -8,7 +8,11 @@
 #' @param csur Include (Under-)Conscript Unicode Registry glyphs.
 #' @param sample Add circle to "Combining" characters.
 #' @inheritParams bittermelon::read_hex
+#' @param cache Use a cached version of this font from `tools::R_user_dir("hexfont", "cache")` if it exists.
+#'              If it does not exist than create a cached version of this font.
 #' @return A [bittermelon::bm_font()] object.
+#'         If `cache` is `TRUE` then as a side effect may create an `.rds` file
+#'         in `tools::R_user_dir("hexfont", "cache")`.
 #' @examples
 #' # Much faster to load only the subset of GNU Unifont one needs
 #' # Mandarin Chinese
@@ -26,9 +30,38 @@
 #'   bm <- as_bm_bitmap(s, font = font)
 #'   print(bm, px = px_ascii)
 #' }
+#'
+#' \donttest{# Will take more than 5s on CRAN machines
+#' # Compiling the entire font from the hex files takes a long time
+#' system.time({font <- unifont()})
+#' length(font) |> prettyNum(big.mark = ",") # number of glyphs
+#  object.size(font) |> format(units = "MB") # memory used
+#' # It is usually much faster to use a cached version of the font
+#' if (file.exists(hexfont:::unifont_cache_filename())) {
+#'   system.time({font_from_cache <- unifont(cache = TRUE)})
+#' }
+#' }
 #' @import bittermelon
 #' @export
-unifont <- function(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE, ucp = NULL) {
+unifont <- function(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE, ucp = NULL, cache = FALSE) {
+    stopifnot(is.null(ucp) || isFALSE(cache))
+    if (cache) {
+        f <- unifont_cache_filename(upper, jp, csur, sample)
+        if (file.exists(f)) {
+            font <- readRDS(f)
+        } else {
+            font <- unifont_from_hex(upper, jp, csur, sample)
+            if (!dir.exists(dirname(f)))
+                dir.create(dirname(f), recursive = TRUE)
+            saveRDS(font, file = f, version = 3)
+        }
+    } else {
+        font <- unifont_from_hex(upper, jp, csur, sample, ucp)
+    }
+    font
+}
+
+unifont_from_hex <- function(upper, jp, csur, sample, ucp = NULL) {
     f_base <- "unifont"
     if (jp) {
         f_base <- paste0(f_base, "_jp")
@@ -58,6 +91,14 @@ unifont <- function(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE, ucp =
     }
     font
 }
+
+unifont_cache_filename <- function(upper = TRUE, jp = FALSE, csur = TRUE, sample = FALSE) {
+    cache_dir <- tools::R_user_dir("hexfont", "cache")
+    filename <- paste0("unifont_", unifont_version(), "_",
+                       sprintf("%d%d%d%d", upper, jp, csur, sample), ".rds")
+    normalizePath(file.path(cache_dir, filename), mustWork = FALSE)
+}
+
 
 #' GNU Unifont version number
 #'